requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.20
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.3
typing-extensions==4.13.2
tzdata==2025.2
